Move Dependency validations into DependencyValidator
